third_party/WebKit/LayoutTests/fast/dom/HTMLObjectElement/update-data.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/js-test.js"></script>
<script>
description("Check that HTMLObjectElement.data updates trigger refresh of element.");

window.jsTestIsAsync = true;
if (window.testRunner)
    testRunner.dumpAsText();

function notify()
{
    testPassed("<object> updated and loaded.");
    finishJSTest();
}

var obj;
var url = "resources/message-ping-on-load.html";
function runTest() {
    obj = document.getElementById("obj");
    shouldBeEqualToString("var u = '" + url + "'; obj.data = u; obj.data.substr(obj.data.indexOf(u))", url);
}
</script>
</head>
<body onmessage="notify()" onload="runTest()">
<object id="obj" type="text/html" data="data:text/html,"></object>
</body>
</html>
