third_party/blink/web_tests/platform/generic/fast/dom/HTMLObjectElement/update-data-expected.txt

Check that HTMLObjectElement.data updates trigger refresh of element.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS var u = 'resources/message-ping-on-load.html'; obj.data = u; obj.data.substr(obj.data.indexOf(u)) is "resources/message-ping-on-load.html"
PASS <object> updated and loaded.
PASS successfullyParsed is true

TEST COMPLETE
